third_party/blink/web_tests/paint/invalidation/compositing/clipped-layer-size-change-expected.txt

CONSOLE MESSAGE: line 41: scrolling
